Tarjeta de Lessor encriptada, Create, edit y delete de Property

+++ Acme-BnB 2.0/Acme-BnB/src/main/java/controllers/lessor/LessorPropertyController.java
@@ -2,9 +2,6 @@
 package controllers.lessor;
 
 import java.util.Collection;
-
-import javax.validation.Valid;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.util.Assert;
@@ -13,12 +10,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
-
-import services.LessorService;
 import services.PropertyService;
 import controllers.AbstractController;
-import domain.Lessor;
 import domain.Property;
+import forms.PropertyForm;
 
 @Controller
 @RequestMapping("/lessor/property")
@@ -29,8 +24,6 @@ public class LessorPropertyController extends AbstractController {
 	@Autowired
 	private PropertyService	propertyService;
 
-	@Autowired
-	private LessorService	lessorService;
 
 
 	//Constructor----------------------
@@ -49,9 +42,9 @@ public class LessorPropertyController extends AbstractController {
 
 		properties = propertyService.findByUserAccount();
 
-		result = new ModelAndView("tenant/list");
+		result = new ModelAndView("property/list");
 		result.addObject("properties", properties);
-		result.addObject("requestURI", "tenant/property/list.do");
+		result.addObject("requestURI", "lessor/property/list.do");
 
 		return result;
 	}
@@ -62,10 +55,10 @@ public class LessorPropertyController extends AbstractController {
 	public ModelAndView create() {
 
 		ModelAndView result;
-		Property property;
+		PropertyForm propertyForm;
 
-		property = propertyService.create();
-		result = createEditModelAndView(property);
+		propertyForm = propertyService.generateForm();
+		result = createEditModelAndView(propertyForm, null);
 
 		return result;
 
@@ -80,27 +73,30 @@ public class LessorPropertyController extends AbstractController {
 		Property property;
 
 		property = propertyService.findOne(propertyId);
+		//PropertyForm propertyForm = propertyService.transform(property);
 		Assert.notNull(property);
-		result = createEditModelAndView(property);
+		result = new ModelAndView("property/edit");
+		result.addObject("property", property);
 
 		return result;
 
 	}
 
 	@RequestMapping(value = "/edit", method = RequestMethod.POST, params = "save")
-	public ModelAndView save(@Valid Property property, BindingResult binding) {
+	public ModelAndView save(Property property, BindingResult binding) {
 
-		ModelAndView result;
-		Lessor lessor = lessorService.findByPrincipal();
-
-		if (binding.hasErrors() || property.getLessor() != lessor) {
-			result = createEditModelAndView(property);
+		ModelAndView result;		
+		property = propertyService.reconstruct(property, binding);
+		
+		if (binding.hasErrors()) {
+			result = list();
 		} else {
 			try {
-				propertyService.save(property);
+				
+				propertyService.save2(property);
 				result = list();
 			} catch (Throwable oops) {
-				result = createEditModelAndView(property, "master.page.commit.error");
+				result =  list();
 			}
 		}
 		return result;
@@ -110,28 +106,34 @@ public class LessorPropertyController extends AbstractController {
 	public ModelAndView delete(Property property, BindingResult binding) {
 
 		ModelAndView result;
-
-		try {
-			propertyService.delete(property);
-			result = list();
-		} catch (Throwable oops) {
-			result = createEditModelAndView(property, "master.page.commit.error");
+		
+		property = propertyService.reconstruct(property, binding);
+		if (binding.hasErrors()) {
+			result = createEditModelAndView(property);
+		} else {
+			try {
+				propertyService.delete(property);
+				result = list();
+			} catch (Throwable oops) {
+				result = createEditModelAndView(property);
+			}
 		}
 		return result;
 	}
 
 	//Ancillary Methods---------------------------
 
-	protected ModelAndView createEditModelAndView(Property property) {
 
+	protected ModelAndView createEditModelAndView(Property property) {
 		ModelAndView result;
 
-		result = createEditModelAndView(property, null);
-
+		result = new ModelAndView("property/edit");
+		result.addObject("property", property);
 		return result;
-	}
 
-	protected ModelAndView createEditModelAndView(Property property, String message) {
+	}
+	
+	protected ModelAndView createEditModelAndView(PropertyForm property, String message) {
 		ModelAndView result;
 
 		result = new ModelAndView("property/edit");
